% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001functions.R
\name{add_variable}
\alias{add_variable}
\title{\code{add_variable} adds columns to the data matrix}
\usage{
add_variable(data = data, dname = "chc", weights = NULL)
}
\arguments{
\item{data}{data.frame, if \code{dname=="chc"} columns "chc_i" has to be in the data.}

\item{dname}{if \code{dname=="chc"} (dummy for chosen alternative) dummy for the choice 
alternative added, if "weights" weights added}

\item{weights}{Matrix with weights to be added to the data}
}
\value{
data.frame
}
\description{
\code{add_variable} adds columns to the data matrix
}
\examples{
chc <- c(1,2,1,4,3,1,4)
data <- data.frame(choice=chc, x=rnorm(length(chc)), y=rnorm(length(chc)))
add_variable(data, dname="chc")
ww <- c(1,1,1,2,2,2,3)
add_variable(data, dname="weights", weights=ww)
}
