% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/006functions.R
\name{bread.nmm}
\alias{bread.nmm}
\title{Bread for Sandwiches.}
\usage{
bread.nmm(object, type = "normal", new_coef = NULL)
}
\arguments{
\item{object}{Fitted \code{nmm} model.}

\item{type}{Type of bread: "normal", "robust", "clustered".}

\item{new_coef}{New coefficients.}
}
\description{
Bread for Sandwiches.
}
\keyword{internal}
