% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/007functions.R
\name{calc_sigma_rho}
\alias{calc_sigma_rho}
\title{Returns function for sigma and rho calculation}
\usage{
calc_sigma_rho(obj)
}
\arguments{
\item{obj}{nmm object}
}
\value{
Function that can be evaluated
}
\description{
Returns function for sigma and rho calculation
}
\keyword{internal}
