% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004functions.R
\name{check_names_start}
\alias{check_names_start}
\title{Checks if starting values have names}
\usage{
check_names_start(start_v)
}
\description{
Checks if starting values have names
}
\keyword{internal}
