% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000functions.R
\name{check_par_av}
\alias{check_par_av}
\title{\code{check_par_av} Checks if all elements of par are available in object.}
\usage{
check_par_av(par, object)
}
\arguments{
\item{par}{String vector, what to look for.}

\item{object}{String vector, where to look.}
}
\description{
\code{check_par_av} Checks if all elements of par are available in object.
}
\keyword{internal}
