% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000functions.R
\name{collapse_space}
\alias{collapse_space}
\title{Collapses lines ending with " "}
\usage{
collapse_space(exx, i, ind)
}
\arguments{
\item{exx}{expression}

\item{i}{first element of ind}

\item{ind}{vector of lines ending with " "}
}
\description{
Collapses lines ending with " "
}
\keyword{internal}
