% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000functions.R
\name{combine_attr_deriv}
\alias{combine_attr_deriv}
\title{\code{combine_attr_deriv} combines attributes of two \code{deriv()} results.}
\usage{
combine_attr_deriv(t1, t2, attribute)
}
\arguments{
\item{t1}{Derivative of first expression. Results of \code{deriv()}.}

\item{t2}{Derivative of second expression. Results of \code{deriv()}.}

\item{attribute}{"grad" for gradient or "hessian" for the Hessian matrix.}
}
\value{
Combined matrices of attributes.
}
\description{
\code{combine_attr_deriv} combines attributes of two \code{deriv()} results.
}
\keyword{internal}
