% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000functions.R
\name{cond_mean_cov_expr}
\alias{cond_mean_cov_expr}
\title{\code{cond_mean_cov_expr} returns conditional mean and variance X|Y of conditional multivariate 
normal distribution . X is discrete variable, Y are continuous variables. 
No correlation between discrete eq.}
\usage{
cond_mean_cov_expr(neqt, kk)
}
\arguments{
\item{neqt}{Number of continuous variables/equations}

\item{kk}{Number of discrete variables/equations}
}
\value{
list:
\item{meannewel}{conditional mean expressions, w/ "<-"}
\item{nomin}{general form of nominator terms, w/o "<-"}
\item{nominl}{for each i in 1:kk nominator terms, w/ "<-"}
\item{denoml}{denominator in conditional mean, w/ "<-"}
\item{denom}{denominator in conditional mean, w/o "<-"}
\item{cel}{conditional covariance expressions, w/ "<-"}
}
\description{
\code{cond_mean_cov_expr} returns conditional mean and variance X|Y of conditional multivariate 
normal distribution . X is discrete variable, Y are continuous variables. 
No correlation between discrete eq.
}
\keyword{internal}
