% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/006functions.R
\name{corstarsl}
\alias{corstarsl}
\title{Significance of correlation matrix}
\usage{
corstarsl(x, xdigit = 2)
}
\arguments{
\item{x}{correlation matrix}
}
\description{
Significance of correlation matrix
}
\keyword{internal}
