% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000generate.R
\docType{data}
\name{dat4cond_mean_cov_expr}
\alias{dat4cond_mean_cov_expr}
\title{Log-likelihood expressions for cont. equations plus 1 discrete}
\format{
An object of class \code{list} of length 4.
}
\usage{
data(dat4cond_mean_cov_expr)
}
\description{
Log-likelihood expressions for cont. equations plus 1 discrete
}
\examples{
data(dat4cond_mean_cov_expr)
}
\keyword{datasets}
