% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000generate.R
\docType{data}
\name{datmaxle}
\alias{datmaxle}
\title{Log-likelihood expressions for cont. equations}
\format{
An object of class \code{list} of length 4.
}
\usage{
data(datmaxle)
}
\description{
Log-likelihood expressions for cont. equations
}
\examples{
data(datmaxle)
}
\keyword{datasets}
