% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000generate.R
\docType{data}
\name{datmlsem}
\alias{datmlsem}
\title{Log-likelihood expressions for cont. equations sem}
\format{
An object of class \code{list} of length 4.
}
\usage{
data(datmlsem)
}
\description{
Log-likelihood expressions for cont. equations sem
}
\examples{
data(datmlsem)
}
\keyword{datasets}
