% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002functions.R
\name{deriv_disc_expr}
\alias{deriv_disc_expr}
\title{Get expression of Hessian and Gradient for discrete choice model with quantile transformation}
\usage{
deriv_disc_expr(
  mnf1 = NULL,
  smodu1,
  attribute = "grad",
  ffor = ffor,
  MNtypef = "logit"
)
}
\arguments{
\item{mnf1}{object containing MNL expressions}

\item{smodu1}{parameter matrix}

\item{attribute}{can be "grad" or "hessian"}

\item{ffor}{discrete choice equations}

\item{MNtypef}{logit or dogit}
}
\value{
derivative expression of discrete equations
}
\description{
Get expression of Hessian and Gradient for discrete choice model with quantile transformation
}
\keyword{internal}
