% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005functions.R
\name{deriv_disc_expr_corr}
\alias{deriv_disc_expr_corr}
\title{Hessian and Gradient expressions}
\usage{
deriv_disc_expr_corr(
  mnf1 = NULL,
  smodu1,
  attribute = "grad",
  ffor = ffor,
  cheqs0 = cheqs0
)
}
\arguments{
\item{mnf1}{object containing MNL expressions}

\item{smodu1}{parameter matrix}

\item{attribute}{can be "grad" or "hessian"}

\item{ffor}{discrete choice equations}

\item{cheqs0}{errors from cont. equations}
}
\value{
derivative expression of discrete equations, with dependency on base mode
}
\description{
Get expression of Hessian and Gradient for discrete choice model with quantile transformation.
}
\keyword{internal}
