% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000generate.R
\docType{data}
\name{expr_ll_norm}
\alias{expr_ll_norm}
\title{Log-likelihood expressions for cont. equations}
\format{
An object of class \code{list} of length 8.
}
\usage{
data(expr_ll_norm)
}
\description{
Log-likelihood expressions for cont. equations
}
\examples{
data(expr_ll_norm)
}
\keyword{datasets}
