% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000generate.R
\docType{data}
\name{expr_ll_norm_v2}
\alias{expr_ll_norm_v2}
\title{Another Log-likelihood expressions for cont. equations version 2}
\format{
An object of class \code{list} of length 8.
}
\usage{
data(expr_ll_norm_v2)
}
\description{
Another Log-likelihood expressions for cont. equations version 2
}
\examples{
data(expr_ll_norm_v2)
}
\keyword{datasets}
