% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/006functions.R
\name{extract_cont_data}
\alias{extract_cont_data}
\title{Takes out only data used in continuous estimation}
\usage{
extract_cont_data(data, cont_eq, par_c)
}
\arguments{
\item{data}{full data matrix}

\item{cont_eq}{cont. equations}

\item{par_c}{parameter vector of cont. equations}
}
\description{
Takes out only data used in continuous estimation
}
\keyword{internal}
