% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001functions.R
\name{f_help}
\alias{f_help}
\title{Internal helper function}
\usage{
f_help(mn, data)
}
\arguments{
\item{mn}{expression to be converted to a function}

\item{data}{name of data frame}
}
\description{
Internal helper function
}
\keyword{internal}
