% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000functions.R
\name{formula2string}
\alias{formula2string}
\title{\code{formula2string} removes square brackets from the supplied expressions.
Convert par[2] -> par2, sigma[2] -> sigma_2_, sigma[2,3] -> sigma_2x2}
\usage{
formula2string(x)
}
\arguments{
\item{x}{String with square brackets.}
}
\value{
String without square brackets.
}
\description{
\code{formula2string} removes square brackets from the supplied expressions.
Convert par[2] -> par2, sigma[2] -> sigma_2_, sigma[2,3] -> sigma_2x2
}
\examples{
xx <- "par[1]*3+par[2]*par+rho1[1,2]+sigma1[2]"
formula2string(xx)
}
