% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001functions.R
\name{generate_disc_coefs}
\alias{generate_disc_coefs}
\title{Function for finding starting values for nls equations with errors}
\usage{
generate_disc_coefs(coefs, eqest, i = 1, type_select = "smallest")
}
\description{
Function for finding starting values for nls equations with errors
}
\keyword{internal}
