% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001functions.R
\name{get_start_cont}
\alias{get_start_cont}
\title{\code{get_start_cont} get starting values for continuous equations.}
\usage{
get_start_cont(
  of,
  hf,
  datcont,
  eq_c,
  par_c,
  startvals = NULL,
  DEoptim_run = TRUE,
  deconst = 2
)
}
\arguments{
\item{of}{Function to me optimized.}

\item{hf}{Hessian function.}

\item{datcont}{Data-set used in the optimization.}

\item{eq_c}{Equations that will be estimated with \code{nlsystemfit}.}

\item{par_c}{Names of parameters.}

\item{startvals}{Supplied starting values.}

\item{DEoptim_run}{if \code{TRUE}, \code{DEoptim} is run for optimization.}

\item{deconst}{Absolute value of lower and upper bounds for \code{DEoptim}.}
}
\value{
Starting values for continuous equations.
}
\description{
\code{get_start_cont} get starting values for continuous equations.
}
\keyword{internal}
