% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001functions.R
\name{get_start_disc}
\alias{get_start_disc}
\title{\code{get_start_disc} get starting values for discrete choice model.}
\usage{
get_start_disc(
  MNf1,
  hff,
  eqest,
  sdat,
  ppardogit = NULL,
  type_select = "smallest"
)
}
\arguments{
\item{MNf1}{Function to me optimized}

\item{hff}{Hessian function}

\item{eqest}{Equations that will be estimated with nls}

\item{sdat}{Data-set used for in the optimization}

\item{ppardogit}{"dogit" parameters, NULL for "logit"}

\item{type_select}{method for filling missing starting values. Default is "smallest"(closest by the index to the left value).
"biggest" takes closest by index to the right value and "random" - samples randomly values.}
}
\value{
Starting values for discrete equations.
}
\description{
\code{get_start_disc} get starting values for discrete choice model.
}
\keyword{internal}
