% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/006functions.R
\name{gradient.nmm}
\alias{gradient.nmm}
\title{Gradient with supplied coefficients}
\usage{
\method{gradient}{nmm}(
  res,
  new_coef = NULL,
  separatenmm = FALSE,
  transform = FALSE,
  methodopt = "NA",
  ind_grad = FALSE
)
}
\arguments{
\item{res}{result from nmm}

\item{new_coef}{coefficients for which logLik should be calculated}

\item{separatenmm}{return separatenmm LL for each equation}

\item{transform}{to do quantile transformation}

\item{methodopt}{"NA" means that automatic algorithm was used in maxLik, if equal to "BHHH" will
return LL for each individual}

\item{ind_grad}{get individual gradient values}
}
\description{
Gradient with supplied coefficients
}
\keyword{internal}
