% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/006functions.R
\name{help_diagnostics}
\alias{help_diagnostics}
\title{Generates objects needed for diagnostics functions}
\usage{
help_diagnostics(x, env = parent.env())
}
\arguments{
\item{x}{Fitted \code{nmm} model.}

\item{env}{To which environment to write the results.}
}
\description{
Generates objects needed for diagnostics functions
}
\keyword{internal}
