% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001functions.R
\name{help_MNdogitf}
\alias{help_MNdogitf}
\title{Internal helper function for MNdogitf}
\usage{
help_MNdogitf(
  exps,
  transform = FALSE,
  separatenmm = FALSE,
  neq = NULL,
  ppardogit = NULL,
  weight_disc = FALSE
)
}
\description{
Internal helper function for MNdogitf
}
\keyword{internal}
