% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001functions.R
\name{help_MNlogitf}
\alias{help_MNlogitf}
\title{Internal helper function for MNlogitf}
\usage{
help_MNlogitf(
  exps,
  transform = FALSE,
  separatenmm = FALSE,
  neq = NULL,
  weight_disc = FALSE
)
}
\description{
Internal helper function for MNlogitf
}
\keyword{internal}
