% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/006functions.R
\name{hessian.nmm}
\alias{hessian.nmm}
\title{Hessian with supplied coefficients}
\usage{
\method{hessian}{nmm}(
  res,
  new_coef = NULL,
  separatenmm = FALSE,
  transform = FALSE,
  methodopt = "NA"
)
}
\arguments{
\item{res}{result from nmm}

\item{new_coef}{coefficients for which logLik should be calculated}

\item{separatenmm}{return separatenmm LL for each equation}

\item{transform}{to do quantile transformation}

\item{methodopt}{"NA" means that automatic algorithm was used in maxLik, if equal to "BHHH" will
return LL for each individual}
}
\description{
Hessian with supplied coefficients
}
\keyword{internal}
