% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/006functions.R
\name{hitRate}
\alias{hitRate}
\title{Calculate Heat rate}
\usage{
hitRate(x, prob_func)
}
\arguments{
\item{x}{Fitted \code{nmm} model.}

\item{prob_func}{function to calculate probabilities.}
}
\description{
Calculate Heat rate
}
\keyword{internal}
