% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/006functions.R
\name{meat.nmm}
\alias{meat.nmm}
\title{Meat for Sandwiches}
\usage{
meat.nmm(object, type = "normal", new_coef = NULL)
}
\arguments{
\item{object}{Fitted \code{nmm} model.}

\item{type}{Type of bread: "normal", "robust", "clustered".}

\item{new_coef}{New coefficients.}
}
\description{
Meat for Sandwiches
}
\keyword{internal}
