% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/007functions.R
\name{mod_function}
\alias{mod_function}
\title{Modifies function to optimize sigma and rho}
\usage{
mod_function(obj)
}
\arguments{
\item{obj}{nmm object}
}
\value{
Function
}
\description{
Modifies function to optimize sigma and rho
}
\keyword{internal}
