% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004functions.R
\name{nmm_1good}
\alias{nmm_1good}
\title{searches for 1st good optimization method between possible maxLik algorithms}
\usage{
nmm_1good(
  joint_func,
  joint_grad,
  joint_hess,
  start_v,
  method = NULL,
  miterlim,
  possible_m,
  methodopt = "BFGS",
  eq_type,
  corrl
)
}
\description{
searches for 1st good optimization method between possible maxLik algorithms
}
\keyword{internal}
