% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004functions.R
\name{nmm_best}
\alias{nmm_best}
\title{searches for best optimization method between possible maxLik algorithms}
\usage{
nmm_best(
  joint_func,
  joint_grad,
  joint_hess,
  start_v,
  method = NULL,
  miterlim,
  possible_m,
  eq_type,
  corrl
)
}
\description{
searches for best optimization method between possible maxLik algorithms
}
\keyword{internal}
