% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004functions.R
\name{nmm_DEoptim}
\alias{nmm_DEoptim}
\title{searches for 1st good optimization method between possible maxLik algorithms and then applies DEoptim}
\usage{
nmm_DEoptim(
  joint_func,
  joint_grad,
  joint_hess,
  start_v = NULL,
  method = NULL,
  miterlim,
  possible_m,
  deconst = 2,
  eq_type,
  corrl
)
}
\description{
searches for 1st good optimization method between possible maxLik algorithms and then applies DEoptim
}
\keyword{internal}
