% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004functions.R
\name{nmm_lastDEoptim}
\alias{nmm_lastDEoptim}
\title{last time try out DEoptim if no functions before produced "good" results}
\usage{
nmm_lastDEoptim(
  joint_func,
  joint_grad,
  joint_hess,
  start_v = NULL,
  method = NULL,
  miterlim,
  possible_m,
  deconst = 2,
  eq_type,
  corrl
)
}
\description{
last time try out DEoptim if no functions before produced "good" results
}
\keyword{internal}
