% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/006functions.R
\name{nrParam}
\alias{nrParam}
\title{\code{nrParam} return number of parameters used in the estimation.}
\usage{
nrParam(x, free = FALSE)
}
\arguments{
\item{x}{Fitted \code{nmm} model.}

\item{free}{Sum or length of parameters.}
}
\description{
\code{nrParam} return number of parameters used in the estimation.
}
\keyword{internal}
