% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/007functions.R
\name{par2sigma}
\alias{par2sigma}
\title{Converts vector of variances and correlations values into a matrix}
\usage{
par2sigma(par, neq, corr_joint, ndisc)
}
\arguments{
\item{par}{Vector of variances and correlations}

\item{neq}{Number of continuous equations}

\item{corr_joint}{Correlation between blocks}

\item{ndisc}{Number of discrete choice alternatives}
}
\value{
Vector
}
\description{
Converts vector of variances and correlations values into a matrix
}
\keyword{internal}
