% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000functions.R
\name{replace_expression}
\alias{replace_expression}
\title{Replaces expressions ".exprXX" with strings}
\usage{
replace_expression(expre, inde)
}
\arguments{
\item{inde}{vector of lines having expression}

\item{exxe}{expression}
}
\description{
Replaces expressions ".exprXX" with strings
}
\keyword{internal}
