% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000functions.R
\name{replace_par_wrap}
\alias{replace_par_wrap}
\title{\code{replace_par_wrap} replace text with other text, wrapper of \code{replace_par}.}
\usage{
replace_par_wrap(repdat, obj)
}
\arguments{
\item{repdat}{data.frame with columns "old" and "new"}

\item{obj}{character string or vector of strings that will be modified.}
}
\value{
Modified string/vector.
}
\description{
\code{replace_par_wrap} replace text with other text, wrapper of \code{replace_par}.
}
\examples{
eq_c <- c("Tw ~ ((((PH) + (tw)) * (ta - Tc + 2) + (1 + (tw)) * (Ec/w - 2/w) -(1 + (PH))) +
sqrt((((PH) + (tw)) * (ta - Tc + 2) + (1 +(tw)) * (Ec/w - 2/w) - (1 + (PH)))^2 - 4 *
 (1 + (PH) + (tw)) *(-(PH) * (ta - Tc + 2) + (1 - (tw) * (ta - Tc + 2)) * (2/w -Ec/w))))/(2 *
  (1 + (PH) + (tw)))",
"Tf1 ~ (th1) * (ta - (((((PH) + (tw)) * (ta - Tc + 2) + (1 + (tw)) *(Ec/w - 2/w) - (1 + (PH))) +
 sqrt((((PH) + (tw)) * (ta -Tc + 2) + (1 + (tw)) * (Ec/w - 2/w) - (1 + (PH)))^2 - 4 *(1 + (PH) +
  (tw)) * (-(PH) * (ta - Tc + 2) + (1 - (tw) *(ta - Tc + 2)) * (2/w - Ec/w))))/(2 * (1 + (PH) +
   (tw)))) -Tc + 2) - 1",
"Ef1 ~ (ph1)/(PH) * (w * (((((PH) + (tw)) * (ta - Tc + 2) + (1 +(tw)) * (Ec/w - 2/w) -
(1 + (PH))) + sqrt((((PH) + (tw)) *(ta - Tc + 2) + (1 + (tw)) * (Ec/w - 2/w) - (1 + (PH)))^2 -4 *
 (1 + (PH) + (tw)) * (-(PH) * (ta - Tc + 2) + (1 - (tw) *(ta - Tc + 2)) * (2/w - Ec/w))))/(2 *
 (1 + (PH) + (tw)))) -Ec + 2) - 1")
parl <- c("tw","PH","th1","ph1")
parll <- paste0("par[", 1:length(parl), "]")
repdat <- data.frame(old=parl, new=parll)
replace_par(2, repdat, eq_c)
replace_par_wrap(repdat, eq_c)
}
