% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/007functions.R
\name{sigma2par}
\alias{sigma2par}
\title{Converts sigma matrix into parameter vector}
\usage{
sigma2par(x, corr_joint)
}
\arguments{
\item{obj}{nmm object}
}
\description{
Converts sigma matrix into parameter vector
}
\keyword{internal}
