% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000functions.R
\name{string2formula}
\alias{string2formula}
\title{\code{string2formula} add square brackets to expressions.
Reverse of formula2string. Convert par[2] <- par2, sigma[2] <- sigma_2_, sigma[2,3] <- sigma_2x2}
\usage{
string2formula(x)
}
\arguments{
\item{x}{String without square brackets.}
}
\value{
String with square brackets.
}
\description{
\code{string2formula} add square brackets to expressions.
Reverse of formula2string. Convert par[2] <- par2, sigma[2] <- sigma_2_, sigma[2,3] <- sigma_2x2
}
\examples{
xx <- "par[1]*3+par[2]*par+rho1[1,2]+sigma1[2]"
xm <- formula2string(xx)
string2formula(xm)
}
