% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004functions.R
\name{try_alg}
\alias{try_alg}
\title{Test if maxLik produces an error}
\usage{
try_alg(
  joint_func,
  joint_grad,
  joint_hess,
  methodim,
  miterlim,
  start_val,
  ki = 1,
  k = 1,
  eq_type,
  corrl = FALSE
)
}
\arguments{
\item{methodim}{- method to try out}

\item{ki}{- iteration number}

\item{k}{- how many times to try}

\item{start_v}{- starting values}
}
\description{
Test if maxLik produces an error
}
\keyword{internal}
