% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004functions.R
\name{try_if}
\alias{try_if}
\title{Checks if nmm object has an error}
\usage{
try_if(res)
}
\arguments{
\item{res}{maxLik object for which summary and other statistics have to be checked}
}
\description{
Checks if nmm object has an error
}
\keyword{internal}
